dashboard/requirements.txt
streamlit
pandas
requests
google-cloud-bigquery